backend/pyproject.toml
[project]
name = "human-rating-platform-backend"
version = "0.1.0"
description = "Backend service for the Human Rating Platform"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.133.0",
  "uvicorn==0.41.0",
  "sqlalchemy==2.0.47",
  "sqlmodel==0.0.37",
  "alembic==1.18.4",
  "greenlet==3.0.3",
  "pydantic==2.12.5",
  "pydantic-settings[toml]==2.12.0",
  "python-multipart==0.0.22",
  "aiofiles==25.1.0",
  "asyncpg==0.31.0",
  "psycopg2-binary==2.9.11",
  "PyJWT==2.8.0",
  "cryptography==42.0.8",
]

[dependency-groups]
dev = [
  "pytest==9.0.2",
  "httpx==0.28.1",
]

[tool.uv]
package = false
